Refactor AISearch into ProductApiActions
DETAILS

Removed the `ProductAiActions` class and integrated its `AISearch` method into `ProductApiActions`. Updated `using` directives to include necessary namespaces, refactored response construction for improved readability, and updated routing for the AISearch endpoint to ensure continued functionality.
